Start trying to attach codegen level info to Firedrake
The idea is that UFL is not an adequate medium for expressing our full
range of code generation information (PyOP2 needs more in particular).z

Therefore we modify Firedrake objects that inherit from UFL to split
them into codegen and data parts.

We use the 'cargo' method for doing so. This is basically composition
instead of inheritance. The key idea here is that if the data objects
attached to the UFL are not immutable, then the data itself does not
need to be. The implications of this are that a 'stripped' form is the
same object as the original form, we do not need to build a new object
from scratch.