WIP Begin to lift form compilation out of WrapperKernelBuilder
The interface we want is Form + KernelInfo -> WrapperKernel. To do this
I need to lift the bit where we call form compilation to out of the
builder class. This is causing many many things to break.

I have also starting doing some magic with subsets so that could be why
things are falling apart.